+++ .github/workflows/validate-readme.yml
@@ -35,7 +35,7 @@ jobs:
           go-version: '^1.17.7'
 
       - name: 'Generate Readme'
-        run: go run scripts/gen-readme.go
+        run: go run scripts/generate.go readme
 
       - name: 'Validate Diff'
         run: |
@@ -43,6 +43,6 @@ jobs:
           if [[ -n "$STATUS" ]]; then
             echo "ERROR: README has not been updated"
             echo ""
-            echo " Run the following command to update the README: go run scripts/gen-readme.go"
+            echo " Run the following command to update the README: go run scripts/generate.go readme"
             exit 1
           fi
