+++ CONTRIBUTING.md
@@ -1,8 +1,26 @@
 # Google GitHub Actions - Example Workflows
 
+## Example Workflows vs Starter Workflows
+
+### Starter Workflows
+
+Starter workflows should be considered _"as simple as is needed for the service"_. This is usually a common scenario with best practices so users can use it off-the-shelf with their applications. They are integrated with the GitHub user interface and are presented to users based on the types of files that exist in their repositories, see the categories property [here](https://github.com/actions/starter-workflows/blob/main/CONTRIBUTING.md).
+
+Additionally, starter workflows are reviewed by the GitHub team and have a published [contributing guide](https://github.com/actions/starter-workflows/blob/main/CONTRIBUTING.md).
+
+### Example Workflows
+
+Can be used to showcase any functionality for a given action. This may include examples for documentation or a blog article and may have highly specific use cases that don't make sense to surface as starter workflows.
+
+#### Example
+
+A good starter workflow for Cloud Run is to build a Docker container for your application, upload it Google Container Registry and then deploy the container Cloud Run. This is a common starting place and has everything needed to start using Cloud Run.
+
+A bad starter workflow for Cloud Run may have user specific logic or custom scripts and implementation steps. This could be good for a specific use case or documentation/blog article, but isn't simple or generic enough for all users to start with.
+
 ## Adding Workflows
 
-New workflows should be bootstrapped with the provided go script: `go run scripts/new-workflow.go action-name/workflow-name`. This will generate the following items:
+New workflows should be bootstrapped with the provided go script: `go run scripts/generate.go workflow action-name/workflow-name`. This will generate the following items:
 
 - A new directory if it does not exist
   - `example-workflows/workflows/action-name`
@@ -24,7 +42,7 @@ New workflows should be bootstrapped with the provided go script: `go run script
 
 ```bash
 # Basic example workflow
-go run scripts/new-workflow.go auth/auth-simple
+go run scripts/generate.go workflow auth/auth-simple
 
 # Folder Structure
 /example-workflows
@@ -37,10 +55,10 @@ go run scripts/new-workflow.go auth/auth-simple
 
 ```bash
 # Starter workflow, default type (deployments)
-go run scripts/new-workflow.go --starter deploy-cloudrun/cloudrun-docker
+go run scripts/generate.go workflow --starter deploy-cloudrun/cloudrun-docker
 
 # Starter workflow, with type
-go run scripts/new-workflow.go --starter --type="automation" deploy-cloudrun/cloudrun-automation
+go run scripts/generate.go workflow --starter --type="automation" deploy-cloudrun/cloudrun-automation
 
 # Folder Structure
 /example-workflows
@@ -58,14 +76,14 @@ go run scripts/new-workflow.go --starter --type="automation" deploy-cloudrun/clo
 - automation
 - ci
 - code-scanning
-- deployments     (default)
+- deployments (default)
 
 ## Gnerate main `README.md`
 
 The main `README.md` file holds references to all the action folders and the workflows they contain. Run the following command to generate an updated `README.md` file based on the `templates/README.tmpl.md` file:
 
 ```bash
-go run scripts/gen-readme.go
+go run scripts/generate.go readme
 ```
 
 ## Pull Request to GitHub Starter Workflows
